Add files via upload
In this project, we will play with data from one of the banks. This data set contains details of a bank's customers. The target variable is a binary variable reflecting the fact whether the customer left the bank (closed his account) or continues to be a customer. So we will create models to accomplish this, and then we will select the best model with the highest accuracy.